refactor(anolisa): rename subscription module and align CLI help
Rename anolisa-core/src/subscription.rs → register.rs and update
the module declaration in lib.rs to match. This aligns the source
file name with the public-facing CLI verb (`anolisa register`).

Reword user-facing help descriptions from implementation details
("token collection", "stop token upload") to product-level framing
("Join/Leave the Agentic OS Co-Build Program"). Users should see
the value proposition, not the internal mechanism.

Why: the previous wording leaked internal terminology into the CLI
surface, making the commands sound like telemetry opt-in rather than
a co-build program enrollment. The file rename eliminates a naming
mismatch (module called "subscription" but CLI says "register").
